+++ crates/core/src/lib.rs
@@ -6,10 +6,13 @@ mod content;
 mod error;
 mod export;
 mod menu;
+mod models;
 mod route;
+mod settings;
 
 pub use content::*;
 pub use error::*;
 pub use export::*;
 pub use menu::*;
 pub use route::*;
+pub use settings::*;